Abort Trap: 6... Abort Trap: 6 Abort Trap: 6 Abort Trap: 6

+++ trunk/bignum/src/utils.c
@@ -82,7 +82,7 @@ bnt bntpop(cbnt str, char* c)
 	*c = str[0];
 	bnt temp_str = (bnt)malloc((strlen(str)-1)*sizeof(char));
 	if (temp_str != NULL) {
-		temp_str = bntsel(str, 1, strlen(str)-1);
+		temp_str = bntsel(str, 1, (int)strlen(str)-1);
 	}
 	else {
 		printf("Crap, malloc in bntpop failed!!\n");
@@ -97,14 +97,14 @@ bnt bntcrop(cbnt str, unsigned int index)
 	bnt temp_str = (bnt)malloc((strlen(str)-1)*sizeof(char));
 	if (temp_str != NULL) {
 		if (index == 0) {
-			temp_str = bntsel(str, index+1, strlen(str)-1);
+			temp_str = bntsel(str, index+1, (int)strlen(str)-1);
 		}
 		else if (index == strlen(str)-1) {
-			temp_str = bntsel(str, 0, strlen(str)-2);
+			temp_str = bntsel(str, 0, (int)strlen(str)-2);
 		}
 		else {
 			temp_str = bntsel(str,0,index-1);
-			strcat(temp_str, bntsel(str,index+1, strlen(str)-1));
+			strcat(temp_str, bntsel(str,index+1, (int)strlen(str)-1));
 		}
 	}
 	else {
